Module fluidex_common::serde

Extra support for misc types in serde.

Example:

use serde::{Serialize, Deserialize};
use fluidex_common::serde::*;
use fluidex_common::Fr;

#[derive(Debug, Serialize, Deserialize)]
struct Message {
    #[serde(with = "HexArray")]
    signature: [u8; 64],
    #[serde(with = "FrBytes")]
    from: Fr,
    #[serde(with = "FrStr")]
    to: Fr,
}

Traits

FrBytes
Helper trait add serde support to Fr using bytes encoding.

FrStr
Helper trait add serde support to Fr using big decimal string literal encoding.

HexArray
Helper trait add serde support to [u8; N] using hex encoding.